backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.9.2
python-dotenv==1.0.1
httpx==0.27.2

# LangGraph / LangChain
langgraph>=0.2.35
langchain-openai>=0.2.3
langchain-core>=0.3.15
openai>=1.52.0

# Firebase / Firestore
firebase-admin==6.5.0

# GitHub
PyGithub==2.4.0

# Serper API (LinkedIn search via Google)
# No extra package needed — uses httpx which is already listed above

# Numerics
numpy==2.1.2

# Jinja2 for email templates
jinja2==3.1.4

# Async support
aiohttp==3.10.9

# MCP server
mcp>=1.0.0
pytest==8.4.2
